Ejercicio2

+++ Ejercicio2.java
@@ -10,7 +10,7 @@ public class Ejercicio2 {
     public static void main (String args []){
         Scanner sc = new Scanner(System.in);
         System.out.print("Dime la cantidad de euros: ");
-        double n = sc.nextInt();
+        double n = sc.nextDouble();
         double pesetas = 166.38;
         
         System.out.print("La cantidad de "+ n + " euros en pesetas, son: " + n*pesetas);
